requirements.txt
gensim
torchtext
nltk
pyemd
